emacs: add a strategy for finding ocamlmerlin executable. Fixes #115.
It adds two customisable variables:
- merlin-command-search-path
- merlin-command-suffixes (typically "", ".native", ".byte" on Linux)

With this, ocamlmerlin is found the following way.
We search for it in merlin-command-search-path, combining it with the
suffixes in merlin-command-suffices. The search path defaults to:
  '("../_build/src"   ;; for using merlin.el directly from a git checkout
    "../../../bin"    ;; installed in <prefix>/share/emacs/site-lisp/
                      ;; (default 'make install' style)
    "../../../../bin" ;; installed in <prefix>/share/<emacs-flavour>/site-lisp/merlin/ (Debian style)
                      ;; or <prefix>/share/emacs/<version>/site-lisp/
    exec-path) ;; PATH

This works also when merlin-command is an absolute path (suffixes are
tried, but the search path has no effect).

This adds a few commodity functions. I would not be surprised if there
were clearer or shorter ways to do all of this in Emacs Lisp.